+++ test/references/index.js
@@ -1,0 +1,2 @@
+require('./basic')
+require('./virtual-listener')
